emg/emg_software/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.49.1",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pure-python-adb>=0.3.0.dev0",
    "pyqtgraph>=0.14.0",
    "pyserial>=3.5",
    "pyside6>=6.10.1",
    "pywavelets>=1.9.0",
    "qdarkstyle>=3.2.3",
    "scipy>=1.16.2",
    "tqdm>=4.67.1",
]
